Refactor the code based on City Building Game tutorial
The structure of the code was a real mess. There are a number of examples
online of state management systems being used in place of the scenes I used,
and they did it in a much cleaner way. So I have replaced my code with the
code they have used, and will continue to refactor the rest of the code
to fit into this structure, which should set me up for continued success in
the future, and much less head scratching as things get more complicated!